services/user_service/requirements.txt
Flask
gunicorn
python-dotenv
sqlalchemy
Flask-SQLAlchemy
psycopg2
flask-cors